Fix some C lint
stringop-overflow in pico_icmp6.c
stringop-truncation in pico_dev_tun.c pico_dev_tap.c
maybe-uninitialized in pico_socket_ll.c pico_dev_radiotest.c
sign-conversion in picoapp.c
(some of this only flagged by gcc when compiling with -O2)

+++ test/picoapp.c
@@ -84,7 +84,7 @@ static pico_time global_pico_tick = 0;
 static void pico_rand_feed(uint32_t feed)
 {
     if (_rand_seed == 0) {
-        _rand_seed = getpid();
+        _rand_seed = (uint32_t) getpid();
     }
     if (!feed)
         return;
